requirements.txt
beautifulsoup4~=4.12.2
colorama==0.4.6
openai==0.27.2
playsound==1.2.2
python-dotenv==1.0.0
pyyaml==6.0
readability-lxml==0.8.1
requests~=2.28.2
tiktoken==0.3.3
gTTS==2.3.1
docker~=6.0.1
duckduckgo-search
google-api-python-client #(https://developers.google.com/custom-search/v1/overview)
pinecone-client==2.2.1

torch~=2.0.0
numpy~=1.24.2
transformers~=4.27.4
nltk~=3.8.1
tensorflow~=2.12.0
scikit-learn~=1.2.2
Markdown~=3.4.3